Update wire plot to plot putative synapses.
Putative synapse locations are recorded when the `syn_distance`
parameter is set in the `[system]` block of the configuration file.
Currently, all locations are stored where there is a distance
equal-or-less than the specified distance in the configuration file.
Hence, putative locations are even recorded in between of dendrites
of the same neuron. Post-hoc filtering can be done by the user.