Add interstitial branching to NeuroMaC.
The idea is to re-activate inactive front at a certain point in the future.
The exact time is implemented in a basic way as "after a specified number of simulation cycles and given a certain probability". The user has to tag a front to be incorporated in a pool of front that can become re-activated.

*CHANGES the front.Front.extend_front API*

TODO: implement better / biologically meaningful re-activation strategies to form interstitial branches.
Likely, add front.interstitial_conditions as a dict to a branch.